Change: rename schema to db
